# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ABA App API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/patients/{patient_id}/pa-services": {
            "get": {
                "tags": [
                    "PA Services"
                ],
                "summary": "Get all PA services for a patient",
                "operationId": "dca1262bb8f49350e75acfa515056707",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of PA services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pa_services": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaService"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "PA Services"
                ],
                "summary": "Create a new PA service",
                "operationId": "83268284fcc6eaa09dd3947211de6ab8",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pa_services",
                                    "cpt",
                                    "n_units",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "pa_services": {
                                        "type": "string",
                                        "example": "Behavioral Analysis"
                                    },
                                    "cpt": {
                                        "type": "string",
                                        "example": "97151"
                                    },
                                    "n_units": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PA service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PA service created successfully"
                                        },
                                        "pa_service": {
                                            "$ref": "#/components/schemas/PaService"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/patients/{patient_id}/pa-services/{id}": {
            "get": {
                "tags": [
                    "PA Services"
                ],
                "summary": "Get a specific PA service",
                "operationId": "de7e896529d5c12cff225030487dc207",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the PA service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PA service details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pa_service": {
                                            "$ref": "#/components/schemas/PaService"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient or PA service not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PA Services"
                ],
                "summary": "Update a PA service",
                "operationId": "da06ce8f6d58d595dbbfdd5b4210508a",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the PA service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pa_services": {
                                        "type": "string",
                                        "example": "Behavioral Analysis"
                                    },
                                    "cpt": {
                                        "type": "string",
                                        "example": "97151"
                                    },
                                    "n_units": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PA service updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PA service updated successfully"
                                        },
                                        "pa_service": {
                                            "$ref": "#/components/schemas/PaService"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient or PA service not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PA Services"
                ],
                "summary": "Delete a PA service",
                "operationId": "9d89d7a343dc2027366f53c22e3f6150",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the PA service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PA service deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PA service deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient or PA service not found"
                    }
                }
            }
        },
        "/api/v2/patients/{patient_id}/pa-services": {
            "get": {
                "tags": [
                    "Admin/PA Services"
                ],
                "summary": "Get paginated PA services list for a patient",
                "operationId": "1f0487852b4957237caac5a3b2abe153",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaginatedPaServiceResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin/PA Services"
                ],
                "summary": "Create a new PA service",
                "operationId": "2a936b7250606d99539722846b315e55",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pa_services",
                                    "cpt",
                                    "n_units",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "pa_services": {
                                        "type": "string",
                                        "example": "Behavioral Analysis"
                                    },
                                    "cpt": {
                                        "type": "string",
                                        "example": "97151"
                                    },
                                    "n_units": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PA service created successfully"
                    }
                }
            }
        },
        "/api/v2/patients/{patient_id}/pa-services/{id}": {
            "get": {
                "tags": [
                    "Admin/PA Services"
                ],
                "summary": "Get a single PA service",
                "operationId": "d3fcce7d0ae49f17eeaa0e537706d6c0",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin/PA Services"
                ],
                "summary": "Update a PA service",
                "operationId": "fe4802b3c5350fdc90663e58f2c57851",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pa_services": {
                                        "type": "string"
                                    },
                                    "cpt": {
                                        "type": "string"
                                    },
                                    "n_units": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PA service updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/PA Services"
                ],
                "summary": "Delete a PA service",
                "operationId": "381f42d90beb0df64630315880b1b1ad",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PA service deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaService": {
                "required": [
                    "patient_id",
                    "pa_services",
                    "cpt",
                    "n_units",
                    "start_date",
                    "end_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "patient_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "pa_services": {
                        "type": "string",
                        "example": "Behavioral Analysis"
                    },
                    "cpt": {
                        "type": "string",
                        "example": "97151"
                    },
                    "n_units": {
                        "type": "integer",
                        "example": 8
                    },
                    "spent_units": {
                        "type": "integer",
                        "example": 0
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-03-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-04-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-03-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-03-01T12:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "datetime",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}